DTSPB-5097: use single tokens

+++ bin/ccd-add-role.sh
@@ -27,16 +27,25 @@ esac
 
 binFolder=$(dirname "$0")
 
-userToken=$(${binFolder}/idam-lease-user-token.sh ${CCD_CONFIGURER_IMPORTER_USERNAME:-ccd.docker.default@hmcts.net} ${CCD_CONFIGURER_IMPORTER_PASSWORD:-Pa55word11})
-serviceToken=$(${binFolder}/idam-lease-service-token.sh ccd_gw $(docker run --rm hmctsprod.azurecr.io/imported/toolbelt/oathtool --totp -b ${API_GATEWAY_S2S_KEY:-AAAAAAAAAAAAAAAA}))
+if [ -z "$USER_TOKEN_ENV" ]; then
+  export USER_TOKEN_ENV=$(${binFolder}/idam-lease-user-token.sh ${CCD_CONFIGURER_IMPORTER_USERNAME:-ccd.docker.default@hmcts.net} ${CCD_CONFIGURER_IMPORTER_PASSWORD:-Pa55word11})
+fi
+
+if [ -z "$SERVICE_TOKEN_ENV" ]; then
+then
+  export SERVICE_TOKEN_ENV=$(${binFolder}/idam-lease-service-token.sh ccd_gw $(docker run --rm hmctsprod.azurecr.io/imported/toolbelt/oathtool --totp -b ${API_GATEWAY_S2S_KEY:-AAAAAAAAAAAAAAAA}))
+fi
+
+
 ccdUrl=${CCD_DEFINITION_STORE_API_BASE_URL:-http://localhost:4451}
 
 echo "Creating CCD role: ${role}"
+}
 
 curl --insecure --fail --show-error --silent --output /dev/null -X PUT \
   ${CCD_DEFINITION_STORE_API_BASE_URL:-http://localhost:4451}/api/user-role \
-  -H "Authorization: Bearer ${userToken}" \
-  -H "ServiceAuthorization: Bearer ${serviceToken}" \
+  -H "Authorization: Bearer ${USER_TOKEN_ENV}" \
+  -H "ServiceAuthorization: Bearer ${SERVICE_TOKEN_ENV}" \
   -H "Content-Type: application/json" \
   -d '{
     "role": "'${role}'",
